DNM: WriteAPIImpl: extend Flush() to flush retry queue
Flush() is documented to force all pending writes from the buffer to
be sent.  It does force pending writes in the buffer to be submitted
for send, but if there is a pending retry with an unexpired timeout it
won't actually send anything.

Make Flush() more useful by having it trigger a re-examination of the
retry queue and initiating a retry immediately.

Depending on whether the documented behavior was intentional, this
operation might be better implemented as an additional WriteAPI method
FlushAll().

Signed-off-by: Peter A. Bigot <pab@pabigot.com>