feat: add the ability to query with SQL
The latest version of InfluxDB Cloud comes with the ability to send SQL
queries. This is done using a new gRPC+TLS client. For now the basic
query function will create the different, non-REST client and return an
Arrow record slice. Users can then go through that slice as necessary.